app/serializers/jwt_token.py
from typing import Optional
from pydantic import BaseModel, Field


class Token(BaseModel):
    access_token: str = Field(alias="accessToken")
    token_type: str = Field(alias="tokenType")


class TokenData(BaseModel):
    username: Optional[str] = None
